Kill flow.master - create master.Master
Also extract events into .events
