+++ mitmproxy/ctx.py
@@ -1,2 +1,2 @@
-master = None  # type: "mitmproxy.flow.FlowMaster"
+master = None  # type: "mitmproxy.master.Master"
 log = None  # type: "mitmproxy.controller.Log"
