+++ mitmproxy/flow/__init__.py
@@ -1,10 +1,9 @@
 from mitmproxy.flow import export
 from mitmproxy.flow.io import FlowWriter, FilteredFlowWriter, FlowReader, read_flows_from_paths
-from mitmproxy.flow.master import FlowMaster
 from mitmproxy.flow.state import State, FlowView
 
 __all__ = [
     "export",
     "FlowWriter", "FilteredFlowWriter", "FlowReader", "read_flows_from_paths",
-    "FlowMaster", "State", "FlowView",
+    "State", "FlowView",
 ]
