+++ mitmproxy/master.py
@@ -1,40 +1,117 @@
 import os
+import threading
+import contextlib
+import queue
 import sys
 
-from netlib import http
+from mitmproxy import addons
+from mitmproxy import options
 from mitmproxy import controller
+from mitmproxy import events
 from mitmproxy import exceptions
 from mitmproxy import models
 from mitmproxy.flow import io
 from mitmproxy.protocol import http_replay
+from netlib import basethread
+from netlib import http
+
+from . import ctx as mitmproxy_ctx
 
 
-def event_sequence(f):
-    if isinstance(f, models.HTTPFlow):
-        if f.request:
-            yield "requestheaders", f
-            yield "request", f
-        if f.response:
-            yield "responseheaders", f
-            yield "response", f
-        if f.error:
-            yield "error", f
-    elif isinstance(f, models.TCPFlow):
-        messages = f.messages
-        f.messages = []
-        f.reply = controller.DummyReply()
-        yield "tcp_start", f
-        while messages:
-            f.messages.append(messages.pop(0))
-            yield "tcp_message", f
-        if f.error:
-            yield "tcp_error", f
-        yield "tcp_end", f
-    else:
-        raise NotImplementedError
+class ServerThread(basethread.BaseThread):
+    def __init__(self, server):
+        self.server = server
+        address = getattr(self.server, "address", None)
+        super().__init__(
+            "ServerThread ({})".format(repr(address))
+        )
+
+    def run(self):
+        self.server.serve_forever()
 
 
-class FlowMaster(controller.Master):
+class Master:
+    """
+        The master handles mitmproxy's main event loop.
+    """
+    def __init__(self, opts, server):
+        self.options = opts or options.Options()
+        self.addons = addons.Addons(self)
+        self.event_queue = queue.Queue()
+        self.should_exit = threading.Event()
+        self.server = server
+        channel = controller.Channel(self.event_queue, self.should_exit)
+        server.set_channel(channel)
+
+    @contextlib.contextmanager
+    def handlecontext(self):
+        # Handlecontexts also have to nest - leave cleanup to the outermost
+        if mitmproxy_ctx.master:
+            yield
+            return
+        mitmproxy_ctx.master = self
+        mitmproxy_ctx.log = controller.Log(self)
+        try:
+            yield
+        finally:
+            mitmproxy_ctx.master = None
+            mitmproxy_ctx.log = None
+
+    def tell(self, mtype, m):
+        m.reply = controller.DummyReply()
+        self.event_queue.put((mtype, m))
+
+    def add_log(self, e, level):
+        """
+            level: debug, info, warn, error
+        """
+        with self.handlecontext():
+            self.addons("log", controller.LogEntry(e, level))
+
+    def start(self):
+        self.should_exit.clear()
+        ServerThread(self.server).start()
+
+    def run(self):
+        self.start()
+        try:
+            while not self.should_exit.is_set():
+                # Don't choose a very small timeout in Python 2:
+                # https://github.com/mitmproxy/mitmproxy/issues/443
+                # TODO: Lower the timeout value if we move to Python 3.
+                self.tick(0.1)
+        finally:
+            self.shutdown()
+
+    def tick(self, timeout):
+        with self.handlecontext():
+            self.addons("tick")
+        changed = False
+        try:
+            mtype, obj = self.event_queue.get(timeout=timeout)
+            if mtype not in events.Events:
+                raise exceptions.ControlException("Unknown event %s" % repr(mtype))
+            handle_func = getattr(self, mtype)
+            if not callable(handle_func):
+                raise exceptions.ControlException("Handler %s not callable" % mtype)
+            if not handle_func.__dict__.get("__handler"):
+                raise exceptions.ControlException(
+                    "Handler function %s is not decorated with controller.handler" % (
+                        handle_func
+                    )
+                )
+            handle_func(obj)
+            self.event_queue.task_done()
+            changed = True
+        except queue.Empty:
+            pass
+        return changed
+
+    def shutdown(self):
+        self.server.shutdown()
+        self.should_exit.set()
+        self.addons.done()
+
     def create_request(self, method, scheme, host, port, path):
         """
             this method creates a new artificial and minimalist request also adds it to flowlist
@@ -70,7 +147,7 @@ class FlowMaster(controller.Master):
                 f.request.port = self.server.config.upstream_server.address.port
                 f.request.scheme = self.server.config.upstream_server.scheme
         f.reply = controller.DummyReply()
-        for e, o in event_sequence(f):
+        for e, o in events.event_sequence(f):
             getattr(self, e)(o)
 
     def load_flows(self, fr):
