+++ test/mitmproxy/builtins/test_anticomp.py
@@ -1,6 +1,6 @@
 from .. import tutils, mastertest
 from mitmproxy.builtins import anticomp
-from mitmproxy.flow import master
+from mitmproxy import master
 from mitmproxy import options
 from mitmproxy import proxy
 
@@ -8,7 +8,7 @@ from mitmproxy import proxy
 class TestAntiComp(mastertest.MasterTest):
     def test_simple(self):
         o = options.Options(anticomp = True)
-        m = master.FlowMaster(o, proxy.DummyServer())
+        m = master.Master(o, proxy.DummyServer())
         sa = anticomp.AntiComp()
         m.addons.add(sa)
