+++ test/mitmproxy/builtins/test_streambodies.py
@@ -1,5 +1,5 @@
 from .. import tutils, mastertest
-from mitmproxy.flow import master
+from mitmproxy import master
 from mitmproxy import options
 from mitmproxy import proxy
 
@@ -9,7 +9,7 @@ from mitmproxy.builtins import streambodies
 class TestStreamBodies(mastertest.MasterTest):
     def test_simple(self):
         o = options.Options(stream_large_bodies = 10)
-        m = master.FlowMaster(o, proxy.DummyServer())
+        m = master.Master(o, proxy.DummyServer())
         sa = streambodies.StreamBodies()
         m.addons.add(sa)
