+++ test/mitmproxy/mastertest.py
@@ -3,7 +3,7 @@ import contextlib
 from . import tutils
 import netlib.tutils
 
-from mitmproxy.flow import master
+from mitmproxy import master
 from mitmproxy import flow, proxy, models, options
 
 
@@ -39,9 +39,9 @@ class MasterTest:
         f.close()
 
 
-class RecordingMaster(master.FlowMaster):
+class RecordingMaster(master.Master):
     def __init__(self, *args, **kwargs):
-        master.FlowMaster.__init__(self, *args, **kwargs)
+        master.Master.__init__(self, *args, **kwargs)
         self.event_log = []
 
     def add_log(self, e, level):
