+++ test/mitmproxy/test_addons.py
@@ -1,6 +1,6 @@
 from mitmproxy import addons
-from mitmproxy import controller
 from mitmproxy import options
+from mitmproxy import master
 from mitmproxy import proxy
 
 
@@ -14,7 +14,7 @@ class TAddon:
 
 def test_simple():
     o = options.Options()
-    m = controller.Master(o, proxy.DummyServer(o))
+    m = master.Master(o, proxy.DummyServer(o))
     a = addons.Addons(m)
     a.add(TAddon("one"))
     assert a.get("one")
